Refactor install actions to separate unit
